+++ src/main/java/bowling/domain/frame/FrameNumber.java
@@ -1,4 +1,4 @@
-package bowling.domain;
+package bowling.domain.frame;
 
 public class FrameNumber implements Comparable<FrameNumber> {
